requirements.txt
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
openai>=1.3.5
python-dotenv
colorama
llama-cpp-python
tqdm